requirements.txt
bcrypt==4.0.1
fastapi==0.116.1
langchain_core==0.3.75
langchain_google_genai==2.1.10
mangum==0.19.0
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
python-multipart==0.0.20
python_jose==3.5.0
SQLAlchemy==2.0.43
uvicorn==0.35.0
wikipedia==1.4.0
